requirements.txt
google-auth-oauthlib==1.2.0
google-auth==2.29.0
google-api-python-client==2.130.0
chromadb==0.5.0
PyMuPDF==1.24.5
python-docx==1.1.0
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.35.0
pandas==2.2.2
plotly==5.20.0
rich==13.7.1

